Split controller.Button into a separate file
